+++ components/ocs_io/i2c/itransceiver.h
@@ -24,7 +24,7 @@ public:
     //!  - @p timeout - interval to wait for the operation to complete,
     //!       -1 means wait forever.
     virtual status::StatusCode
-    send(const uint8_t* buf, unsigned size, system::Time timeout) = 0;
+    send(const uint8_t* buf, size_t size, system::Time timeout) = 0;
 
     //! Receive data from the I2C device.
     //!
@@ -33,7 +33,7 @@ public:
     //!  - @p timeout - interval to wait for the operation to complete,
     //!       -1 means wait forever.
     virtual status::StatusCode
-    receive(uint8_t* buf, unsigned size, system::Time timeout) = 0;
+    receive(uint8_t* buf, size_t size, system::Time timeout) = 0;
 };
 
 } // namespace i2c
