+++ components/ocs_onewire/serial_number_to_str.h
@@ -16,7 +16,7 @@ namespace onewire {
 //! Format sensor serial number to string.
 class serial_number_to_str : private core::NonCopyable<> {
 public:
-    static constexpr unsigned str_length = strlen("AA:BB:CC:DD:EE:FF");
+    static constexpr size_t str_length = strlen("AA:BB:CC:DD:EE:FF");
 
     //! Initialize.
     serial_number_to_str(const SerialNumber& serial_number);
